HBX-1640: Move class 'org.hibernate.tool.hbm2x.DocExporter' to 'org.hibernate.tool.internal.export.doc.DocExporter'
