+++ orm/src/main/java/org/hibernate/tool/ant/Hbm2DocExporterTask.java
@@ -5,7 +5,7 @@
 package org.hibernate.tool.ant;
 
 import org.hibernate.tool.api.export.Exporter;
-import org.hibernate.tool.hbm2x.DocExporter;
+import org.hibernate.tool.internal.export.doc.DocExporter;
 
 public class Hbm2DocExporterTask extends ExporterTask {
